[CSCI 3130] Assignment 3 - Email Validator Stage 2
Added functionalities to account for whitespace and alpha-numeric
characters in address. Refactored check from stage 1 to simply check
that the address suffix was allowable (for the purposes of this
assignment, I stuck with a .com address).

Modified validate() function to iterate over a series of regexes as
discussed above and validate in stages. Note that this is NOT the
most efficient way of doing this, but after research I found that the
regex for a valid email is incredibly complex, and since I did not
want to rip the regex from someone else, I simply crafted it to check
in increments.